chore: combine api and proxy into one cli command
